Move helper functions to separate file
resolver helper functions can be kept in a separate file so that neat_resolver
is a bit more manageable.
