+++ neat_internal.h
@@ -274,9 +274,6 @@ uint8_t neat_resolve(struct neat_resolver *resolver,
                          uint16_t port,
                          neat_resolver_handle_t handle_resolve,
                          void *user_data);
-//Check if node is an IP literal or not. Returns -1 on failure, 0 if not
-//literal, 1 if literal
-int8_t neat_resolver_check_for_literal(uint8_t *family, const char *node);
 
 //Update timeouts (in ms) for DNS resolving. T1 is total timeout, T2 is how long
 //to wait after first reply from DNS server. Initial values are 30s and 1s.
